+++ src/co/edu/upb/nominae/pojos/EVO.java
@@ -1,11 +1,11 @@
 package co.edu.upb.nominae.pojos;
 
+import java.math.BigDecimal;
 import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlTransient;
 import javax.xml.bind.annotation.XmlType;
 
-//@XmlRootElement(name = "EVO")
+
 @XmlType (propOrder={"HZRNEVO_DOTACION",
     "HZRNEVO_APOYO_SOSTENIMIENT",
     "HZRNEVO_TELETRABAJO",
@@ -16,65 +16,65 @@ import javax.xml.bind.annotation.XmlType;
 public class EVO {
     
     @XmlElement(name = "EVO_1")
-    private Double HZRNEVO_DOTACION;
+    private BigDecimal HZRNEVO_DOTACION;
     @XmlElement(name = "EVO_2")
-    private Double HZRNEVO_APOYO_SOSTENIMIENT;
+    private BigDecimal HZRNEVO_APOYO_SOSTENIMIENT;
     @XmlElement(name = "EVO_3")
-    private Double HZRNEVO_TELETRABAJO;
+    private BigDecimal HZRNEVO_TELETRABAJO;
     @XmlElement(name = "EVO_4")
-    private Double HZRNEVO_BONIF_RETIRO;
+    private BigDecimal HZRNEVO_BONIF_RETIRO;
     @XmlElement(name = "EVO_5")
-    private Double HZRNEVO_INDEMNIZACION;
+    private BigDecimal HZRNEVO_INDEMNIZACION;
     @XmlElement(name = "EVO_6")
-    private Double HZRNEVO_REINTEGRO;    
+    private BigDecimal HZRNEVO_REINTEGRO;    
     private Integer HZRNEVO_ITE_ID;
 
     @XmlTransient
-    public Double getHZRNEVO_DOTACION() {
+    public BigDecimal getHZRNEVO_DOTACION() {
         return HZRNEVO_DOTACION;
     }
 
-    public void setHZRNEVO_DOTACION(Double HZRNEVO_DOTACION) {
+    public void setHZRNEVO_DOTACION(BigDecimal HZRNEVO_DOTACION) {
         this.HZRNEVO_DOTACION = HZRNEVO_DOTACION;
     }
     @XmlTransient
-    public Double getHZRNEVO_APOYO_SOSTENIMIENT() {
+    public BigDecimal getHZRNEVO_APOYO_SOSTENIMIENT() {
         return HZRNEVO_APOYO_SOSTENIMIENT;
     }
 
-    public void setHZRNEVO_APOYO_SOSTENIMIENT(Double HZRNEVO_APOYO_SOSTENIMIENT) {
+    public void setHZRNEVO_APOYO_SOSTENIMIENT(BigDecimal HZRNEVO_APOYO_SOSTENIMIENT) {
         this.HZRNEVO_APOYO_SOSTENIMIENT = HZRNEVO_APOYO_SOSTENIMIENT;
     }
     @XmlTransient
-    public Double getHZRNEVO_TELETRABAJO() {
+    public BigDecimal getHZRNEVO_TELETRABAJO() {
         return HZRNEVO_TELETRABAJO;
     }
 
-    public void setHZRNEVO_TELETRABAJO(Double HZRNEVO_TELETRABAJO) {
+    public void setHZRNEVO_TELETRABAJO(BigDecimal HZRNEVO_TELETRABAJO) {
         this.HZRNEVO_TELETRABAJO = HZRNEVO_TELETRABAJO;
     }
     @XmlTransient
-    public Double getHZRNEVO_BONIF_RETIRO() {
+    public BigDecimal getHZRNEVO_BONIF_RETIRO() {
         return HZRNEVO_BONIF_RETIRO;
     }
 
-    public void setHZRNEVO_BONIF_RETIRO(Double HZRNEVO_BONIF_RETIRO) {
+    public void setHZRNEVO_BONIF_RETIRO(BigDecimal HZRNEVO_BONIF_RETIRO) {
         this.HZRNEVO_BONIF_RETIRO = HZRNEVO_BONIF_RETIRO;
     }
     @XmlTransient
-    public Double getHZRNEVO_INDEMNIZACION() {
+    public BigDecimal getHZRNEVO_INDEMNIZACION() {
         return HZRNEVO_INDEMNIZACION;
     }
 
-    public void setHZRNEVO_INDEMNIZACION(Double HZRNEVO_INDEMNIZACION) {
+    public void setHZRNEVO_INDEMNIZACION(BigDecimal HZRNEVO_INDEMNIZACION) {
         this.HZRNEVO_INDEMNIZACION = HZRNEVO_INDEMNIZACION;
     }
     @XmlTransient
-    public Double getHZRNEVO_REINTEGRO() {
+    public BigDecimal getHZRNEVO_REINTEGRO() {
         return HZRNEVO_REINTEGRO;
     }
 
-    public void setHZRNEVO_REINTEGRO(Double HZRNEVO_REINTEGRO) {
+    public void setHZRNEVO_REINTEGRO(BigDecimal HZRNEVO_REINTEGRO) {
         this.HZRNEVO_REINTEGRO = HZRNEVO_REINTEGRO;
     }
     @XmlTransient
