+++ src/co/edu/upb/nominae/pojos/SAN.java
@@ -1,29 +1,29 @@
 package co.edu.upb.nominae.pojos;
 
 import co.edu.upb.utilities.Data;
+import java.math.BigDecimal;
 import java.util.zip.DataFormatException;
 import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlTransient;
 import javax.xml.bind.annotation.XmlType;
 
-//@XmlRootElement(name = "SAN")
+
 @XmlType (propOrder={"HZRNSAN_SANCION_PUBLICA",
     "HZRNSAN_SANCION_PRIVADA"})
 public class SAN {
     
     @XmlElement(name = "SAN_1")
-    private Double HZRNSAN_SANCION_PUBLICA;
+    private BigDecimal HZRNSAN_SANCION_PUBLICA;
     @XmlElement(name = "SAN_2")
-    private Double HZRNSAN_SANCION_PRIVADA;
+    private BigDecimal HZRNSAN_SANCION_PRIVADA;
     private Integer HZRNSAN_ITS_ID;       
 
     @XmlTransient
-    public Double getHZRNSAN_SANCION_PUBLICA() {
+    public BigDecimal getHZRNSAN_SANCION_PUBLICA() {
         return HZRNSAN_SANCION_PUBLICA;
     }
 
-    public void setHZRNSAN_SANCION_PUBLICA(Double HZRNSAN_SANCION_PUBLICA) throws DataFormatException {
+    public void setHZRNSAN_SANCION_PUBLICA(BigDecimal HZRNSAN_SANCION_PUBLICA) throws DataFormatException {
         Data data = new Data();    
 
         if(data.checklength(HZRNSAN_SANCION_PUBLICA,15,6)){
@@ -34,11 +34,11 @@ public class SAN {
         }          
     }
     @XmlTransient
-    public Double getHZRNSAN_SANCION_PRIVADA() {
+    public BigDecimal getHZRNSAN_SANCION_PRIVADA() {
         return HZRNSAN_SANCION_PRIVADA;
     }
 
-    public void setHZRNSAN_SANCION_PRIVADA(Double HZRNSAN_SANCION_PRIVADA) throws DataFormatException {
+    public void setHZRNSAN_SANCION_PRIVADA(BigDecimal HZRNSAN_SANCION_PRIVADA) throws DataFormatException {
         Data data = new Data();    
 
         if(data.checklength(HZRNSAN_SANCION_PRIVADA,15,6)){
