+++ src/co/edu/upb/nominae/pojos/SPE.java
@@ -1,30 +1,30 @@
 package co.edu.upb.nominae.pojos;
 
 import co.edu.upb.utilities.Data;
+import java.math.BigDecimal;
 import java.util.zip.DataFormatException;
 import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
+
 import javax.xml.bind.annotation.XmlTransient;
 import javax.xml.bind.annotation.XmlType;
 
-//@XmlRootElement(name = "SPE")
+
 @XmlType (propOrder={"HZRNSPE_PORCENTAJE",
     "HZRNSPE_DEDUCCION"})
-
 public class SPE {
     
     @XmlElement(name = "SPE_1")
-    private Double HZRNSPE_PORCENTAJE;
+    private BigDecimal HZRNSPE_PORCENTAJE;
     @XmlElement(name = "SPE_2")
-    private Double HZRNSPE_DEDUCCION;
+    private BigDecimal HZRNSPE_DEDUCCION;
     private Integer HZRNSPE_ITS_ID;      
 
     @XmlTransient
-    public Double getHZRNSPE_PORCENTAJE() {
+    public BigDecimal getHZRNSPE_PORCENTAJE() {
         return HZRNSPE_PORCENTAJE;
     }
 
-    public void setHZRNSPE_PORCENTAJE(Double HZRNSPE_PORCENTAJE) throws DataFormatException {
+    public void setHZRNSPE_PORCENTAJE(BigDecimal HZRNSPE_PORCENTAJE) throws DataFormatException {
         Data data = new Data();    
 
         if(data.checklength(HZRNSPE_PORCENTAJE,3,2)){
@@ -35,11 +35,11 @@ public class SPE {
         }          
     }
     @XmlTransient
-    public Double getHZRNSPE_DEDUCCION() {
+    public BigDecimal getHZRNSPE_DEDUCCION() {
         return HZRNSPE_DEDUCCION;
     }
 
-    public void setHZRNSPE_DEDUCCION(Double HZRNSPE_DEDUCCION) throws DataFormatException {
+    public void setHZRNSPE_DEDUCCION(BigDecimal HZRNSPE_DEDUCCION) throws DataFormatException {
         Data data = new Data();    
 
         if(data.checklength(HZRNSPE_DEDUCCION,15,6)){
